fix(packaging): move the host launch table into the shipped tree
`prepack-check` failed: `_lib/headless_invocation.ts` imported HOST_INVENTORY
and HOST_CONFIGS from `src/cli/python/`, which is NOT in the package.json
`files` whitelist. Both resolve fine in a checkout and would have crashed a
global install with ERR_MODULE_NOT_FOUND — so `--print-relaunch` was broken for
every consumer and worked for me.

Nothing local catches this. ESLint, typecheck and the pre-push static pass are
all green on the import, because they only ever see the checkout, where the
path exists.

Two available fixes. Shipping `src/cli/python/` means adding 396 KB and 14
modules — `workspace_secrets` and `workspace_crypto` among them — to the
consumer package so that two constants resolve; widening that surface to fix an
import is the wrong direction. So the ~40 lines the shipped side needs moved
instead, into `_lib/host_launch.ts`:

- `HOST_INVENTORY` — re-exported from `workspace_hosts.ts`, so every importer
  and the ADR-068 contract test are untouched.
- `LAUNCH_ARGV` — the three Tier-1 `build_args`, now the one source
  `HOST_CONFIGS` builds on.

The dependency is INVERTED, not duplicated: the shipped module owns the facts
and the unshipped driver imports them back. Re-deriving them on the shipped
side is exactly the defect round 2 found in this module, where a hand-built
copy modelled codex and gemini as stdin consumers. A new test asserts the drive
loop and the shared table still produce identical argv — it may reach across
the packaging boundary that the production module must not.

Only the declarative half moved. The spawn, the timeout, resume argv and the
per-host envelope parsers stay with the driver, because they have a runtime and
a table does not.

prepack-check green, typecheck green, 142 tests across the five affected suites
including both relocated modules' own.
